🏗️ Mangling with the overall measurement & tore `object.ts` to pieces
