+++ test/unit/object/unit.test.ts
@@ -1,0 +1,8 @@
+import {measurement, unit} from "!src/object";
+
+describe('unit', () => {
+  it("should return the unit of the measurement.", () => {
+    const actual: "px" = unit(measurement(12.5, "px"));
+    expect(actual).toBe("px");
+  });
+});
